使用主控端的网页窗口大小

+++ Source/ImGui_WS/Private/Incppect/incppect.cpp
@@ -500,8 +500,8 @@ struct Incppect<SSL>::Impl {
 
     Parameters parameters;
 
-    int32_t txTotal_bytes = 0;
-    int32_t rxTotal_bytes = 0;
+    double txTotal_bytes = 0;
+    double rxTotal_bytes = 0;
 
     std::map<TPath, int> pathToGetter;
     std::vector<TGetter> getters;
